src/requirements.txt
elasticsearch==8.0.0
Flask==1.1.2
Jinja2==2.11.3
MarkupSafe==1.1.1
Werkzeug==1.0.1
itsdangerous==1.1.0
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
uuid==1.30
selenium==4.1.5
Flask-BasicAuth==0.2.0
Flask-HTTPBasicAuth==1.0.1
pandas==1.4.2
python-dateutil==2.8.1
beautifulsoup4==4.9.3
azure-storage-blob==12.10.0b1
lxml==4.5.1

